+++ zone/lua_spell.h
@@ -6,12 +6,6 @@
 
 struct SPDat_Spell_Struct;
 
-namespace luabind {
-	struct scope;
-}
-
-luabind::scope lua_register_spell();
-
 class Lua_Spell : public Lua_Ptr<const SPDat_Spell_Struct>
 {
 	typedef const SPDat_Spell_Struct NativeType;
