+++ zone/lua_stat_bonuses_impl.cpp
@@ -1,0 +1,1282 @@
+#include "lua_stat_bonuses.h"
+
+int32 Lua_StatBonuses::GetAC() const {
+	Lua_Safe_Call_Int();
+	return self->AC;
+}
+
+int64 Lua_StatBonuses::GetHP() const {
+	Lua_Safe_Call_Int();
+	return self->HP;
+}
+
+int64 Lua_StatBonuses::GetHPRegen() const {
+	Lua_Safe_Call_Int();
+	return self->HPRegen;
+}
+
+int64 Lua_StatBonuses::GetMaxHP() const {
+	Lua_Safe_Call_Int();
+	return self->MaxHP;
+}
+
+int64 Lua_StatBonuses::GetManaRegen() const {
+	Lua_Safe_Call_Int();
+	return self->ManaRegen;
+}
+
+int64 Lua_StatBonuses::GetEnduranceRegen() const {
+	Lua_Safe_Call_Int();
+	return self->EnduranceRegen;
+}
+
+int64 Lua_StatBonuses::GetMana() const {
+	Lua_Safe_Call_Int();
+	return self->Mana;
+}
+
+int64 Lua_StatBonuses::GetEndurance() const {
+	Lua_Safe_Call_Int();
+	return self->Endurance;
+}
+
+int32 Lua_StatBonuses::GetATK() const {
+	Lua_Safe_Call_Int();
+	return self->ATK;
+}
+
+int32 Lua_StatBonuses::GetSTR() const {
+	Lua_Safe_Call_Int();
+	return self->STR;
+}
+
+int32 Lua_StatBonuses::GetSTRCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->STRCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicSTR() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicSTR;
+}
+
+int32 Lua_StatBonuses::GetSTA() const {
+	Lua_Safe_Call_Int();
+	return self->STA;
+}
+
+int32 Lua_StatBonuses::GetSTACapMod() const {
+	Lua_Safe_Call_Int();
+	return self->STACapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicSTA() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicSTA;
+}
+
+int32 Lua_StatBonuses::GetDEX() const {
+	Lua_Safe_Call_Int();
+	return self->DEX;
+}
+
+int32 Lua_StatBonuses::GetDEXCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->DEXCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicDEX() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicDEX;
+}
+
+int32 Lua_StatBonuses::GetAGI() const {
+	Lua_Safe_Call_Int();
+	return self->AGI;
+}
+
+int32 Lua_StatBonuses::GetAGICapMod() const {
+	Lua_Safe_Call_Int();
+	return self->AGICapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicAGI() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicAGI;
+}
+
+int32 Lua_StatBonuses::GetINT() const {
+	Lua_Safe_Call_Int();
+	return self->INT;
+}
+
+int32 Lua_StatBonuses::GetINTCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->INTCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicINT() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicINT;
+}
+
+int32 Lua_StatBonuses::GetWIS() const {
+	Lua_Safe_Call_Int();
+	return self->WIS;
+}
+
+int32 Lua_StatBonuses::GetWISCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->WISCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicWIS() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicWIS;
+}
+
+int32 Lua_StatBonuses::GetCHA() const {
+	Lua_Safe_Call_Int();
+	return self->CHA;
+}
+
+int32 Lua_StatBonuses::GetCHACapMod() const {
+	Lua_Safe_Call_Int();
+	return self->CHACapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicCHA() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicCHA;
+}
+
+int32 Lua_StatBonuses::GetMR() const {
+	Lua_Safe_Call_Int();
+	return self->MR;
+}
+
+int32 Lua_StatBonuses::GetMRCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->MRCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicMR() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicMR;
+}
+
+int32 Lua_StatBonuses::GetFR() const {
+	Lua_Safe_Call_Int();
+	return self->FR;
+}
+
+int32 Lua_StatBonuses::GetFRCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->FRCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicFR() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicFR;
+}
+
+int32 Lua_StatBonuses::GetCR() const {
+	Lua_Safe_Call_Int();
+	return self->CR;
+}
+
+int32 Lua_StatBonuses::GetCRCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->CRCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicCR() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicCR;
+}
+
+int32 Lua_StatBonuses::GetPR() const {
+	Lua_Safe_Call_Int();
+	return self->PR;
+}
+
+int32 Lua_StatBonuses::GetPRCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->PRCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicPR() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicPR;
+}
+
+int32 Lua_StatBonuses::GetDR() const {
+	Lua_Safe_Call_Int();
+	return self->DR;
+}
+
+int32 Lua_StatBonuses::GetDRCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->DRCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicDR() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicDR;
+}
+
+int32 Lua_StatBonuses::GetCorrup() const {
+	Lua_Safe_Call_Int();
+	return self->Corrup;
+}
+
+int32 Lua_StatBonuses::GetCorrupCapMod() const {
+	Lua_Safe_Call_Int();
+	return self->CorrupCapMod;
+}
+
+int32 Lua_StatBonuses::GetHeroicCorrup() const {
+	Lua_Safe_Call_Int();
+	return self->HeroicCorrup;
+}
+
+uint16 Lua_StatBonuses::GetDamageShieldSpellID() const {
+	Lua_Safe_Call_Int();
+	return self->DamageShieldSpellID;
+}
+
+int Lua_StatBonuses::GetDamageShield() const {
+	Lua_Safe_Call_Int();
+	return self->DamageShield;
+}
+
+int Lua_StatBonuses::GetDamageShieldType() const {
+	Lua_Safe_Call_Int();
+	return self->DamageShieldType;
+}
+
+int Lua_StatBonuses::GetSpellDamageShield() const {
+	Lua_Safe_Call_Int();
+	return self->SpellDamageShield;
+}
+
+int Lua_StatBonuses::GetSpellShield() const {
+	Lua_Safe_Call_Int();
+	return self->SpellShield;
+}
+
+int Lua_StatBonuses::GetReverseDamageShield() const {
+	Lua_Safe_Call_Int();
+	return self->ReverseDamageShield;
+}
+
+uint16 Lua_StatBonuses::GetReverseDamageShieldSpellID() const {
+	Lua_Safe_Call_Int();
+	return self->ReverseDamageShieldSpellID;
+}
+
+int Lua_StatBonuses::GetReverseDamageShieldType() const {
+	Lua_Safe_Call_Int();
+	return self->ReverseDamageShieldType;
+}
+
+int Lua_StatBonuses::Getmovementspeed() const {
+	Lua_Safe_Call_Int();
+	return self->movementspeed;
+}
+
+int32 Lua_StatBonuses::Gethaste() const {
+	Lua_Safe_Call_Int();
+	return self->haste;
+}
+
+int32 Lua_StatBonuses::Gethastetype2() const {
+	Lua_Safe_Call_Int();
+	return self->hastetype2;
+}
+
+int32 Lua_StatBonuses::Gethastetype3() const {
+	Lua_Safe_Call_Int();
+	return self->hastetype3;
+}
+
+int32 Lua_StatBonuses::Getinhibitmelee() const {
+	Lua_Safe_Call_Int();
+	return self->inhibitmelee;
+}
+
+float Lua_StatBonuses::GetAggroRange() const {
+	Lua_Safe_Call_Real();
+	return self->AggroRange;
+}
+
+float Lua_StatBonuses::GetAssistRange() const {
+	Lua_Safe_Call_Real();
+	return self->AssistRange;
+}
+
+int32 Lua_StatBonuses::Getskillmod(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->skillmod[idx];
+}
+
+int32 Lua_StatBonuses::Getskillmodmax(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->skillmodmax[idx];
+}
+
+int Lua_StatBonuses::Geteffective_casting_level() const {
+	Lua_Safe_Call_Int();
+	return self->effective_casting_level;
+}
+
+int Lua_StatBonuses::Getadjusted_casting_skill() const {
+	Lua_Safe_Call_Int();
+	return self->adjusted_casting_skill;
+}
+
+int Lua_StatBonuses::Getreflect_chance() const {
+	Lua_Safe_Call_Int();
+	return self->reflect[SBIndex::REFLECT_CHANCE];
+}
+
+uint32 Lua_StatBonuses::GetsingingMod() const {
+	Lua_Safe_Call_Int();
+	return self->singingMod;
+}
+
+uint32 Lua_StatBonuses::GetAmplification() const {
+	Lua_Safe_Call_Int();
+	return self->Amplification;
+}
+
+uint32 Lua_StatBonuses::GetbrassMod() const {
+	Lua_Safe_Call_Int();
+	return self->brassMod;
+}
+
+uint32 Lua_StatBonuses::GetpercussionMod() const {
+	Lua_Safe_Call_Int();
+	return self->percussionMod;
+}
+
+uint32 Lua_StatBonuses::GetwindMod() const {
+	Lua_Safe_Call_Int();
+	return self->windMod;
+}
+
+uint32 Lua_StatBonuses::GetstringedMod() const {
+	Lua_Safe_Call_Int();
+	return self->stringedMod;
+}
+
+uint32 Lua_StatBonuses::GetsongModCap() const {
+	Lua_Safe_Call_Int();
+	return self->songModCap;
+}
+
+int32 Lua_StatBonuses::Gethatemod() const {
+	Lua_Safe_Call_Int();
+	return self->hatemod;
+}
+
+int64 Lua_StatBonuses::GetEnduranceReduction() const {
+	Lua_Safe_Call_Int();
+	return self->EnduranceReduction;
+}
+
+int32 Lua_StatBonuses::GetStrikeThrough() const {
+	Lua_Safe_Call_Int();
+	return self->StrikeThrough;
+}
+
+int32 Lua_StatBonuses::GetMeleeMitigation() const {
+	Lua_Safe_Call_Int();
+	return self->MeleeMitigation;
+}
+
+int32 Lua_StatBonuses::GetMeleeMitigationEffect() const {
+	Lua_Safe_Call_Int();
+	return self->MeleeMitigationEffect;
+}
+
+int32 Lua_StatBonuses::GetCriticalHitChance(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->CriticalHitChance[idx];
+}
+
+int32 Lua_StatBonuses::GetCriticalSpellChance() const {
+	Lua_Safe_Call_Int();
+	return self->CriticalSpellChance;
+}
+
+int32 Lua_StatBonuses::GetSpellCritDmgIncrease() const {
+	Lua_Safe_Call_Int();
+	return self->SpellCritDmgIncrease;
+}
+
+int32 Lua_StatBonuses::GetSpellCritDmgIncNoStack() const {
+	Lua_Safe_Call_Int();
+	return self->SpellCritDmgIncNoStack;
+}
+
+int32 Lua_StatBonuses::GetDotCritDmgIncrease() const {
+	Lua_Safe_Call_Int();
+	return self->DotCritDmgIncrease;
+}
+
+int32 Lua_StatBonuses::GetCriticalHealChance() const {
+	Lua_Safe_Call_Int();
+	return self->CriticalHealChance;
+}
+
+int32 Lua_StatBonuses::GetCriticalHealOverTime() const {
+	Lua_Safe_Call_Int();
+	return self->CriticalHealOverTime;
+}
+
+int32 Lua_StatBonuses::GetCriticalDoTChance() const {
+	Lua_Safe_Call_Int();
+	return self->CriticalDoTChance;
+}
+
+int32 Lua_StatBonuses::GetCrippBlowChance() const {
+	Lua_Safe_Call_Int();
+	return self->CrippBlowChance;
+}
+
+int32 Lua_StatBonuses::GetAvoidMeleeChance() const {
+	Lua_Safe_Call_Int();
+	return self->AvoidMeleeChance;
+}
+
+int32 Lua_StatBonuses::GetAvoidMeleeChanceEffect() const {
+	Lua_Safe_Call_Int();
+	return self->AvoidMeleeChanceEffect;
+}
+
+int32 Lua_StatBonuses::GetRiposteChance() const {
+	Lua_Safe_Call_Int();
+	return self->RiposteChance;
+}
+
+int32 Lua_StatBonuses::GetDodgeChance() const {
+	Lua_Safe_Call_Int();
+	return self->DodgeChance;
+}
+
+int32 Lua_StatBonuses::GetParryChance() const {
+	Lua_Safe_Call_Int();
+	return self->ParryChance;
+}
+
+int32 Lua_StatBonuses::GetDualWieldChance() const {
+	Lua_Safe_Call_Int();
+	return self->DualWieldChance;
+}
+
+int32 Lua_StatBonuses::GetDoubleAttackChance() const {
+	Lua_Safe_Call_Int();
+	return self->DoubleAttackChance;
+}
+
+int32 Lua_StatBonuses::GetTripleAttackChance() const {
+	Lua_Safe_Call_Int();
+	return self->TripleAttackChance;
+}
+
+int32 Lua_StatBonuses::GetDoubleRangedAttack() const {
+	Lua_Safe_Call_Int();
+	return self->DoubleRangedAttack;
+}
+
+int32 Lua_StatBonuses::GetResistSpellChance() const {
+	Lua_Safe_Call_Int();
+	return self->ResistSpellChance;
+}
+
+int32 Lua_StatBonuses::GetResistFearChance() const {
+	Lua_Safe_Call_Int();
+	return self->ResistFearChance;
+}
+
+bool Lua_StatBonuses::GetFearless() const {
+	Lua_Safe_Call_Bool();
+	return self->Fearless;
+}
+
+bool Lua_StatBonuses::GetIsFeared() const {
+	Lua_Safe_Call_Bool();
+	return self->IsFeared;
+}
+
+bool Lua_StatBonuses::GetIsBlind() const {
+	Lua_Safe_Call_Bool();
+	return self->IsBlind;
+}
+
+int32 Lua_StatBonuses::GetStunResist() const {
+	Lua_Safe_Call_Int();
+	return self->StunResist;
+}
+
+int32 Lua_StatBonuses::GetMeleeSkillCheck() const {
+	Lua_Safe_Call_Int();
+	return self->MeleeSkillCheck;
+}
+
+uint8 Lua_StatBonuses::GetMeleeSkillCheckSkill() const {
+	Lua_Safe_Call_Int();
+	return self->MeleeSkillCheckSkill;
+}
+
+int32 Lua_StatBonuses::GetHitChance() const {
+	Lua_Safe_Call_Int();
+	return self->HitChance;
+}
+
+int32 Lua_StatBonuses::GetHitChanceEffect(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->HitChanceEffect[idx];
+}
+
+int32 Lua_StatBonuses::GetDamageModifier(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->DamageModifier[idx];
+}
+
+int32 Lua_StatBonuses::GetDamageModifier2(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->DamageModifier2[idx];
+}
+
+int32 Lua_StatBonuses::GetMinDamageModifier(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->MinDamageModifier[idx];
+}
+
+int32 Lua_StatBonuses::GetProcChance() const {
+	Lua_Safe_Call_Int();
+	return self->ProcChance;
+}
+
+int32 Lua_StatBonuses::GetProcChanceSPA() const {
+	Lua_Safe_Call_Int();
+	return self->ProcChanceSPA;
+}
+
+int32 Lua_StatBonuses::GetExtraAttackChance() const {
+	Lua_Safe_Call_Int();
+	return self->ExtraAttackChance[0];
+}
+
+int32 Lua_StatBonuses::GetDoTShielding() const {
+	Lua_Safe_Call_Int();
+	return self->DoTShielding;
+}
+
+int32 Lua_StatBonuses::GetFlurryChance() const {
+	Lua_Safe_Call_Int();
+	return self->FlurryChance;
+}
+
+int32 Lua_StatBonuses::GetHundredHands() const {
+	Lua_Safe_Call_Int();
+	return self->HundredHands;
+}
+
+int32 Lua_StatBonuses::GetMeleeLifetap() const {
+	Lua_Safe_Call_Int();
+	return self->MeleeLifetap;
+}
+
+int32 Lua_StatBonuses::GetVampirism() const {
+	Lua_Safe_Call_Int();
+	return self->Vampirism;
+}
+
+int32 Lua_StatBonuses::GetHealRate() const {
+	Lua_Safe_Call_Int();
+	return self->HealRate;
+}
+
+int32 Lua_StatBonuses::GetMaxHPChange() const {
+	Lua_Safe_Call_Int();
+	return self->MaxHPChange;
+}
+
+int32 Lua_StatBonuses::GetHealAmt() const {
+	Lua_Safe_Call_Int();
+	return self->HealAmt;
+}
+
+int32 Lua_StatBonuses::GetSpellDmg() const {
+	Lua_Safe_Call_Int();
+	return self->SpellDmg;
+}
+
+int32 Lua_StatBonuses::GetClairvoyance() const {
+	Lua_Safe_Call_Int();
+	return self->Clairvoyance;
+}
+
+int32 Lua_StatBonuses::GetDSMitigation() const {
+	Lua_Safe_Call_Int();
+	return self->DSMitigation;
+}
+
+int32 Lua_StatBonuses::GetDSMitigationOffHand() const {
+	Lua_Safe_Call_Int();
+	return self->DSMitigationOffHand;
+}
+
+int32 Lua_StatBonuses::GetTwoHandBluntBlock() const {
+	Lua_Safe_Call_Int();
+	return self->TwoHandBluntBlock;
+}
+
+uint32 Lua_StatBonuses::GetItemManaRegenCap() const {
+	Lua_Safe_Call_Int();
+	return self->ItemManaRegenCap;
+}
+
+int32 Lua_StatBonuses::GetGravityEffect() const {
+	Lua_Safe_Call_Int();
+	return self->GravityEffect;
+}
+
+bool Lua_StatBonuses::GetAntiGate() const {
+	Lua_Safe_Call_Bool();
+	return self->AntiGate;
+}
+
+bool Lua_StatBonuses::GetMagicWeapon() const {
+	Lua_Safe_Call_Bool();
+	return self->MagicWeapon;
+}
+
+int32 Lua_StatBonuses::GetIncreaseBlockChance() const {
+	Lua_Safe_Call_Int();
+	return self->IncreaseBlockChance;
+}
+
+uint32 Lua_StatBonuses::GetPersistantCasting() const {
+	Lua_Safe_Call_Int();
+	return self->PersistantCasting;
+}
+
+int Lua_StatBonuses::GetXPRateMod() const {
+	Lua_Safe_Call_Int();
+	return self->XPRateMod;
+}
+
+bool Lua_StatBonuses::GetBlockNextSpell() const {
+	Lua_Safe_Call_Bool();
+	return false; // Bonus no longer used due to effect being a focus
+}
+
+bool Lua_StatBonuses::GetImmuneToFlee() const {
+	Lua_Safe_Call_Bool();
+	return self->ImmuneToFlee;
+}
+
+uint32 Lua_StatBonuses::GetVoiceGraft() const {
+	Lua_Safe_Call_Int();
+	return self->VoiceGraft;
+}
+
+int32 Lua_StatBonuses::GetSpellProcChance() const {
+	Lua_Safe_Call_Int();
+	return self->SpellProcChance;
+}
+
+int32 Lua_StatBonuses::GetCharmBreakChance() const {
+	Lua_Safe_Call_Int();
+	return self->CharmBreakChance;
+}
+
+int32 Lua_StatBonuses::GetSongRange() const {
+	Lua_Safe_Call_Int();
+	return self->SongRange;
+}
+
+uint32 Lua_StatBonuses::GetHPToManaConvert() const {
+	Lua_Safe_Call_Int();
+	return self->HPToManaConvert;
+}
+
+bool Lua_StatBonuses::GetNegateEffects() const {
+	Lua_Safe_Call_Bool();
+	return self->NegateEffects;
+}
+
+bool Lua_StatBonuses::GetTriggerMeleeThreshold() const {
+	Lua_Safe_Call_Bool();
+	return self->TriggerMeleeThreshold;
+}
+
+bool Lua_StatBonuses::GetTriggerSpellThreshold() const {
+	Lua_Safe_Call_Bool();
+	return self->TriggerSpellThreshold;
+}
+
+int32 Lua_StatBonuses::GetShieldBlock() const {
+	Lua_Safe_Call_Int();
+	return self->ShieldBlock;
+}
+
+int32 Lua_StatBonuses::GetBlockBehind() const {
+	Lua_Safe_Call_Int();
+	return self->BlockBehind;
+}
+
+bool Lua_StatBonuses::GetCriticalRegenDecay() const {
+	Lua_Safe_Call_Bool();
+	return self->CriticalRegenDecay;
+}
+
+bool Lua_StatBonuses::GetCriticalHealDecay() const {
+	Lua_Safe_Call_Bool();
+	return self->CriticalHealDecay;
+}
+
+bool Lua_StatBonuses::GetCriticalDotDecay() const {
+	Lua_Safe_Call_Bool();
+	return self->CriticalDotDecay;
+}
+
+bool Lua_StatBonuses::GetDivineAura() const {
+	Lua_Safe_Call_Bool();
+	return self->DivineAura;
+}
+
+bool Lua_StatBonuses::GetDistanceRemoval() const {
+	Lua_Safe_Call_Bool();
+	return self->DistanceRemoval;
+}
+
+int32 Lua_StatBonuses::GetFrenziedDevastation() const {
+	Lua_Safe_Call_Int();
+	return self->FrenziedDevastation;
+}
+
+bool Lua_StatBonuses::GetNegateIfCombat() const {
+	Lua_Safe_Call_Bool();
+	return self->NegateIfCombat;
+}
+
+int8 Lua_StatBonuses::GetScreech() const {
+	Lua_Safe_Call_Int();
+	return self->Screech;
+}
+
+int32 Lua_StatBonuses::GetAlterNPCLevel() const {
+	Lua_Safe_Call_Int();
+	return self->AlterNPCLevel;
+}
+
+bool Lua_StatBonuses::GetBerserkSPA() const {
+	Lua_Safe_Call_Bool();
+	return self->BerserkSPA;
+}
+
+int32 Lua_StatBonuses::GetMetabolism() const {
+	Lua_Safe_Call_Int();
+	return self->Metabolism;
+}
+
+bool Lua_StatBonuses::GetSanctuary() const {
+	Lua_Safe_Call_Bool();
+	return self->Sanctuary;
+}
+
+int32 Lua_StatBonuses::GetFactionModPct() const {
+	Lua_Safe_Call_Int();
+	return self->FactionModPct;
+}
+
+uint32 Lua_StatBonuses::GetPC_Pet_Flurry() const {
+	Lua_Safe_Call_Int();
+	return self->PC_Pet_Flurry;
+}
+
+int8 Lua_StatBonuses::GetPackrat() const {
+	Lua_Safe_Call_Int();
+	return self->Packrat;
+}
+
+uint8 Lua_StatBonuses::GetBuffSlotIncrease() const {
+	Lua_Safe_Call_Int();
+	return self->BuffSlotIncrease;
+}
+
+uint32 Lua_StatBonuses::GetDelayDeath() const {
+	Lua_Safe_Call_Int();
+	return self->DelayDeath;
+}
+
+int8 Lua_StatBonuses::GetBaseMovementSpeed() const {
+	Lua_Safe_Call_Int();
+	return self->BaseMovementSpeed;
+}
+
+uint8 Lua_StatBonuses::GetIncreaseRunSpeedCap() const {
+	Lua_Safe_Call_Int();
+	return self->IncreaseRunSpeedCap;
+}
+
+int32 Lua_StatBonuses::GetDoubleSpecialAttack() const {
+	Lua_Safe_Call_Int();
+	return self->DoubleSpecialAttack;
+}
+
+uint8 Lua_StatBonuses::GetFrontalStunResist() const {
+	Lua_Safe_Call_Int();
+	return self->FrontalStunResist;
+}
+
+int32 Lua_StatBonuses::GetBindWound() const {
+	Lua_Safe_Call_Int();
+	return self->BindWound;
+}
+
+int32 Lua_StatBonuses::GetMaxBindWound() const {
+	Lua_Safe_Call_Int();
+	return self->MaxBindWound;
+}
+
+int32 Lua_StatBonuses::GetChannelChanceSpells() const {
+	Lua_Safe_Call_Int();
+	return self->ChannelChanceSpells;
+}
+
+int32 Lua_StatBonuses::GetChannelChanceItems() const {
+	Lua_Safe_Call_Int();
+	return self->ChannelChanceItems;
+}
+
+uint8 Lua_StatBonuses::GetSeeInvis() const {
+	Lua_Safe_Call_Int();
+	return self->SeeInvis;
+}
+
+uint8 Lua_StatBonuses::GetTripleBackstab() const {
+	Lua_Safe_Call_Int();
+	return self->TripleBackstab;
+}
+
+bool Lua_StatBonuses::GetFrontalBackstabMinDmg() const {
+	Lua_Safe_Call_Bool();
+	return self->FrontalBackstabMinDmg;
+}
+
+uint8 Lua_StatBonuses::GetFrontalBackstabChance() const {
+	Lua_Safe_Call_Int();
+	return self->FrontalBackstabChance;
+}
+
+uint8 Lua_StatBonuses::GetConsumeProjectile() const {
+	Lua_Safe_Call_Int();
+	return self->ConsumeProjectile;
+}
+
+uint8 Lua_StatBonuses::GetForageAdditionalItems() const {
+	Lua_Safe_Call_Int();
+	return self->ForageAdditionalItems;
+}
+
+uint8 Lua_StatBonuses::GetSalvageChance() const {
+	Lua_Safe_Call_Int();
+	return self->SalvageChance;
+}
+
+uint32 Lua_StatBonuses::GetArcheryDamageModifier() const {
+	Lua_Safe_Call_Int();
+	return self->ArcheryDamageModifier;
+}
+
+bool Lua_StatBonuses::GetSecondaryDmgInc() const {
+	Lua_Safe_Call_Bool();
+	return self->SecondaryDmgInc;
+}
+
+uint32 Lua_StatBonuses::GetGiveDoubleAttack() const {
+	Lua_Safe_Call_Int();
+	return self->GiveDoubleAttack;
+}
+
+int32 Lua_StatBonuses::GetPetCriticalHit() const {
+	Lua_Safe_Call_Int();
+	return self->PetCriticalHit;
+}
+
+int32 Lua_StatBonuses::GetPetAvoidance() const {
+	Lua_Safe_Call_Int();
+	return self->PetAvoidance;
+}
+
+int32 Lua_StatBonuses::GetCombatStability() const {
+	Lua_Safe_Call_Int();
+	return self->CombatStability;
+}
+
+int32 Lua_StatBonuses::GetDoubleRiposte() const {
+	Lua_Safe_Call_Int();
+	return self->DoubleRiposte;
+}
+
+int32 Lua_StatBonuses::GetAmbidexterity() const {
+	Lua_Safe_Call_Int();
+	return self->Ambidexterity;
+}
+
+int32 Lua_StatBonuses::GetPetMaxHP() const {
+	Lua_Safe_Call_Int();
+	return self->PetMaxHP;
+}
+
+int32 Lua_StatBonuses::GetPetFlurry() const {
+	Lua_Safe_Call_Int();
+	return self->PetFlurry;
+}
+
+uint8 Lua_StatBonuses::GetMasteryofPast() const {
+	Lua_Safe_Call_Int();
+	return self->MasteryofPast;
+}
+
+bool Lua_StatBonuses::GetGivePetGroupTarget() const {
+	Lua_Safe_Call_Bool();
+	return self->GivePetGroupTarget;
+}
+
+int32 Lua_StatBonuses::GetRootBreakChance() const {
+	Lua_Safe_Call_Int();
+	return self->RootBreakChance;
+}
+
+int32 Lua_StatBonuses::GetUnfailingDivinity() const {
+	Lua_Safe_Call_Int();
+	return self->UnfailingDivinity;
+}
+
+int32 Lua_StatBonuses::GetItemHPRegenCap() const {
+	Lua_Safe_Call_Int();
+	return self->ItemHPRegenCap;
+}
+
+int32 Lua_StatBonuses::GetOffhandRiposteFail() const {
+	Lua_Safe_Call_Int();
+	return self->OffhandRiposteFail;
+}
+
+int32 Lua_StatBonuses::GetItemATKCap() const {
+	Lua_Safe_Call_Int();
+	return self->ItemATKCap;
+}
+
+int32 Lua_StatBonuses::GetShieldEquipDmgMod() const {
+	Lua_Safe_Call_Int();
+	return self->ShieldEquipDmgMod;
+}
+
+bool Lua_StatBonuses::GetTriggerOnValueAmount() const {
+	Lua_Safe_Call_Bool();
+	return self->TriggerOnCastRequirement;
+}
+
+int8 Lua_StatBonuses::GetStunBashChance() const {
+	Lua_Safe_Call_Int();
+	return self->StunBashChance;
+}
+
+int8 Lua_StatBonuses::GetIncreaseChanceMemwipe() const {
+	Lua_Safe_Call_Int();
+	return self->IncreaseChanceMemwipe;
+}
+
+int8 Lua_StatBonuses::GetCriticalMend() const {
+	Lua_Safe_Call_Int();
+	return self->CriticalMend;
+}
+
+int32 Lua_StatBonuses::GetImprovedReclaimEnergy() const {
+	Lua_Safe_Call_Int();
+	return self->ImprovedReclaimEnergy;
+}
+
+int32 Lua_StatBonuses::GetPetMeleeMitigation() const {
+	Lua_Safe_Call_Int();
+	return self->PetMeleeMitigation;
+}
+
+bool Lua_StatBonuses::GetIllusionPersistence() const {
+	Lua_Safe_Call_Bool();
+	return self->IllusionPersistence;
+}
+
+uint16 Lua_StatBonuses::Getextra_xtargets() const {
+	Lua_Safe_Call_Int();
+	return self->extra_xtargets;
+}
+
+bool Lua_StatBonuses::GetShroudofStealth() const {
+	Lua_Safe_Call_Bool();
+	return self->ShroudofStealth;
+}
+
+uint16 Lua_StatBonuses::GetReduceFallDamage() const {
+	Lua_Safe_Call_Int();
+	return self->ReduceFallDamage;
+}
+
+uint8 Lua_StatBonuses::GetTradeSkillMastery() const {
+	Lua_Safe_Call_Int();
+	return self->TradeSkillMastery;
+}
+
+int16 Lua_StatBonuses::GetNoBreakAESneak() const {
+	Lua_Safe_Call_Int();
+	return self->NoBreakAESneak;
+}
+
+int16 Lua_StatBonuses::GetFeignedCastOnChance() const {
+	Lua_Safe_Call_Int();
+	return self->FeignedCastOnChance;
+}
+
+int32 Lua_StatBonuses::GetDivineSaveChance(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->DivineSaveChance[idx];
+}
+
+uint32 Lua_StatBonuses::GetDeathSave(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->DeathSave[idx];
+}
+
+int32 Lua_StatBonuses::GetAccuracy(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->Accuracy[idx];
+}
+
+int16 Lua_StatBonuses::GetSkillDmgTaken(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SkillDmgTaken[idx];
+}
+
+uint32 Lua_StatBonuses::GetSpellTriggers(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SpellTriggers[idx];
+}
+
+uint32 Lua_StatBonuses::GetSpellOnKill(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SpellOnKill[idx];
+}
+
+uint32 Lua_StatBonuses::GetSpellOnDeath(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SpellOnDeath[idx];
+}
+
+int32 Lua_StatBonuses::GetCritDmgMod(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->CritDmgMod[idx];
+}
+
+int32 Lua_StatBonuses::GetSkillReuseTime(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SkillReuseTime[idx];
+}
+
+int32 Lua_StatBonuses::GetSkillDamageAmount(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SkillDamageAmount[idx];
+}
+
+int Lua_StatBonuses::GetHPPercCap(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->HPPercCap[idx];
+}
+
+int Lua_StatBonuses::GetManaPercCap(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->ManaPercCap[idx];
+}
+
+int Lua_StatBonuses::GetEndPercCap(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->EndPercCap[idx];
+}
+
+uint8 Lua_StatBonuses::GetFocusEffects(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->FocusEffects[idx];
+}
+
+int16 Lua_StatBonuses::GetFocusEffectsWorn(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->FocusEffectsWorn[idx];
+}
+
+int32 Lua_StatBonuses::GetSkillDamageAmount2(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SkillDamageAmount2[idx];
+}
+
+uint32 Lua_StatBonuses::GetNegateAttacks(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->NegateAttacks[idx];
+}
+
+uint32 Lua_StatBonuses::GetMitigateMeleeRune(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->MitigateMeleeRune[idx];
+}
+
+uint32 Lua_StatBonuses::GetMeleeThresholdGuard(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->MeleeThresholdGuard[idx];
+}
+
+uint32 Lua_StatBonuses::GetSpellThresholdGuard(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SpellThresholdGuard[idx];
+}
+
+uint32 Lua_StatBonuses::GetMitigateSpellRune(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->MitigateSpellRune[idx];
+}
+
+uint32 Lua_StatBonuses::GetMitigateDotRune(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->MitigateDotRune[idx];
+}
+
+uint32 Lua_StatBonuses::GetManaAbsorbPercentDamage(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->ManaAbsorbPercentDamage;
+}
+
+int32 Lua_StatBonuses::GetImprovedTaunt(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->ImprovedTaunt[idx];
+}
+
+int8 Lua_StatBonuses::GetRoot(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->Root[idx];
+}
+
+uint32 Lua_StatBonuses::GetAbsorbMagicAtt(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->AbsorbMagicAtt[idx];
+}
+
+uint32 Lua_StatBonuses::GetMeleeRune(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->MeleeRune[idx];
+}
+
+int32 Lua_StatBonuses::GetAStacker(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->AStacker[idx];
+}
+
+int32 Lua_StatBonuses::GetBStacker(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->BStacker[idx];
+}
+
+int32 Lua_StatBonuses::GetCStacker(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->CStacker[idx];
+}
+
+int32 Lua_StatBonuses::GetDStacker(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->DStacker[idx];
+}
+
+bool Lua_StatBonuses::GetLimitToSkill(int idx) const {
+	Lua_Safe_Call_Bool();
+	return self->LimitToSkill[idx];
+}
+
+uint32 Lua_StatBonuses::GetSkillProc(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SkillProc[idx];
+}
+
+uint32 Lua_StatBonuses::GetSkillProcSuccess(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SkillProcSuccess[idx];
+}
+
+uint32 Lua_StatBonuses::GetPC_Pet_Rampage(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->PC_Pet_Rampage[idx];
+}
+
+int32 Lua_StatBonuses::GetSkillAttackProc(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SkillAttackProc[idx];
+}
+
+int32 Lua_StatBonuses::GetSlayUndead(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SlayUndead[idx];
+}
+
+int32 Lua_StatBonuses::GetGiveDoubleRiposte(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->GiveDoubleRiposte[idx];
+}
+
+uint32 Lua_StatBonuses::GetRaiseSkillCap(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->RaiseSkillCap[idx];
+}
+
+int32 Lua_StatBonuses::GetSEResist(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->SEResist[idx];
+}
+
+int32 Lua_StatBonuses::GetFinishingBlow(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->FinishingBlow[idx];
+}
+
+uint32 Lua_StatBonuses::GetFinishingBlowLvl(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->FinishingBlowLvl[idx];
+}
+
+uint32 Lua_StatBonuses::GetHeadShot(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->HeadShot[idx];
+}
+
+uint8 Lua_StatBonuses::GetHSLevel(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->HSLevel[idx];
+}
+
+uint32 Lua_StatBonuses::GetAssassinate(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->Assassinate[idx];
+}
+
+uint8 Lua_StatBonuses::GetAssassinateLevel(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->AssassinateLevel[idx];
+}
+
+int32 Lua_StatBonuses::GetReduceTradeskillFail(int idx) const {
+	Lua_Safe_Call_Int();
+	return self->ReduceTradeskillFail[idx];
+}
+
